backend/requirements.txt
fastapi
uvicorn
neo4j
groq
pandas
openpyxl
python-multipart
pydantic-settings